Reorder functions and add some comments
