+++ x-pack/metricbeat/module/oracle/performance/_meta/cursor_by_username_and_machine_data.json
@@ -11,9 +11,9 @@
     "oracle": {
         "performance": {
             "cursors": {
-                "avg": 0.7073170731707317,
+                "avg": 0.6829268292682927,
                 "max": 17,
-                "total": 29
+                "total": 28
             },
             "machine": "2ed9ac3a4c3d",
             "username": "Unknown"
