+++ x-pack/metricbeat/module/oracle/performance/_meta/data.json
@@ -12,23 +12,23 @@
         "performance": {
             "cursors": {
                 "cache_hit": {
-                    "pct": 0.8843277645186953
+                    "pct": 0.8311065930874237
                 },
                 "opened": {
                     "current": 5,
-                    "total": 6285
+                    "total": 6886
                 },
                 "parse": {
-                    "real": 1872,
-                    "total": 3686
+                    "real": 1524,
+                    "total": 4199
                 },
                 "session": {
-                    "cache_hits": 5558
+                    "cache_hits": 5723
                 }
             },
-            "io_reloads": 0.0018835335983681005,
-            "lock_requests": 0.5878733266651177,
-            "pin_requests": 0.8006816024796614
+            "io_reloads": 0.0015732643391367942,
+            "lock_requests": 0.5958019475347308,
+            "pin_requests": 0.8058170730696832
         }
     },
     "service": {
