+++ x-pack/metricbeat/module/oracle/performance/buffer_cache_hit_ratio.go
@@ -7,6 +7,8 @@ package performance
 import (
 	"context"
 	"database/sql"
+	"github.com/elastic/beats/libbeat/common"
+	"github.com/elastic/beats/x-pack/metricbeat/module/oracle"
 
 	"github.com/pkg/errors"
 )
@@ -20,13 +22,13 @@ type bufferCacheHitRatio struct {
 }
 
 /*
-	The following function executes a query that produces the following result
-
-	NAME	PHYSICAL_READS	DB_BLOCK_GETS	CONSISTENT_GETS	Hit Ratio
-	DEFAULT	19024			17195			379538			0.9520483549389639883750532473981241792339
-
-	Each instance of bufferCacheHitRatio represents a row from the previous results
-*/
+ * The following function executes a query that produces the following result
+ *
+ * NAME	PHYSICAL_READS	DB_BLOCK_GETS	CONSISTENT_GETS	Hit Ratio
+ * DEFAULT	19024			17195			379538			0.9520483549389639883750532473981241792339
+ *
+ * Each instance of bufferCacheHitRatio represents a row from the previous results
+ */
 func (e *performanceExtractor) bufferCacheHitRatio(ctx context.Context) ([]bufferCacheHitRatio, error) {
 	rows, err := e.db.QueryContext(ctx, `SELECT name, physical_reads, db_block_gets, consistent_gets,
        1 - (physical_reads / (db_block_gets + consistent_gets)) "Hit Ratio"
@@ -47,4 +49,25 @@ FROM V$BUFFER_POOL_STATISTICS`)
 	}
 
 	return results, nil
+}
+
+// addTempFreeSpaceData is specific to the TEMP Tablespace.
+func (m *MetricSet) addBufferCacheRatioData(bs []bufferCacheHitRatio) map[string]common.MapStr {
+	out := make(map[string]common.MapStr)
+
+	for _, bufferCacheHitRatio := range bs {
+		if _, found := out[bufferCacheHitRatio.name.String]; !found {
+			out[bufferCacheHitRatio.name.String] = common.MapStr{}
+		}
+
+		_, _ = out[bufferCacheHitRatio.name.String].Put("buffer_pool", bufferCacheHitRatio.name.String)
+
+		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.buffer.hit.pct", &oracle.Float64Value{NullFloat64: bufferCacheHitRatio.hitRatio})
+		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.get.consistent", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.consistentGets})
+		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.get.db_blocks", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.dbBlockGets})
+		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.physical_reads", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.physicalReads})
+
+	}
+
+	return out
 }
