+++ x-pack/metricbeat/module/oracle/performance/data.go
@@ -7,16 +7,13 @@ package performance
 import (
 	"context"
 
-	"github.com/elastic/beats/x-pack/metricbeat/module/oracle"
 
 	"github.com/pkg/errors"
 
-	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/metricbeat/mb"
 )
 
-// extract is the E of a ETL processing. Gets the data files, used/free space and temp free space data that is fetch
-// by doing queries to Oracle
+// extract is the E of a ETL processing. Gets all the performance information into an instance of extractedData
 func (m *MetricSet) extract(ctx context.Context, extractor performanceExtractMethods) (out *extractedData, err error) {
 	out = &extractedData{}
 
@@ -39,6 +36,8 @@ func (m *MetricSet) extract(ctx context.Context, extractor performanceExtractMet
 	return
 }
 
+// extractAndTransform just composes the ET operations from a ETL. It's called by the Fetch method, which is the one
+// that "loads" the data into Elasticsearch
 func (m *MetricSet) extractAndTransform(ctx context.Context) ([]mb.Event, error) {
 	extractedMetricsData, err := m.extract(ctx, m.extractor)
 	if err != nil {
@@ -50,13 +49,13 @@ func (m *MetricSet) extractAndTransform(ctx context.Context) ([]mb.Event, error)
 
 // transform is the T of an ETL (refer to the 'extract' method above if you need to see the origin). Transforms the data
 // to create a Kibana/Elasticsearch friendly JSON. Data from Oracle is pretty fragmented by design so a lot of data
-// was necessary. Data is organized by Tablespace entity (Tablespaces might contain one or more data files)
+// was necessary. More than one different event is generated. Refer to the _meta folder too see ones.
 func (m *MetricSet) transform(in *extractedData) []mb.Event {
 	bufferCache := m.addBufferCacheRatioData(in.bufferCacheHitRatios)
 	cursorByUsernameAndMachineEvents := m.addCursorByUsernameAndMachine(in.cursorsByUsernameAndMachine)
 
 	cursorEvent := m.addCursorData(in.totalCursors)
-	cursorEvent.Update(m.addLibraryData(in.libraryData))
+	cursorEvent.Update(m.addLibraryCacheData(in.libraryData))
 
 	events := make([]mb.Event, 0)
 
@@ -71,68 +70,4 @@ func (m *MetricSet) transform(in *extractedData) []mb.Event {
 	}
 
 	return events
-}
-
-func (m *MetricSet) addCursorData(cs *totalCursors) common.MapStr {
-	out := make(common.MapStr)
-
-	oracle.SetSqlValue(m.Logger(), out, "cursors.opened.total", &oracle.Int64Value{NullInt64: cs.totalCursors})
-	oracle.SetSqlValue(m.Logger(), out, "cursors.opened.current", &oracle.Int64Value{NullInt64: cs.currentCursors})
-	oracle.SetSqlValue(m.Logger(), out, "cursors.session.cache_hits", &oracle.Int64Value{NullInt64: cs.sessCurCacheHits})
-	oracle.SetSqlValue(m.Logger(), out, "cursors.parse.total", &oracle.Int64Value{NullInt64: cs.parseCountTotal})
-	oracle.SetSqlValue(m.Logger(), out, "cursors.cache_hit.pct", &oracle.Float64Value{NullFloat64: cs.cacheHitsTotalCursorsRatio})
-	oracle.SetSqlValue(m.Logger(), out, "cursors.parse.real", &oracle.Int64Value{NullInt64: cs.realParses})
-
-	return out
-}
-
-func (m *MetricSet) addCursorByUsernameAndMachine(cs []cursorsByUsernameAndMachine) []common.MapStr {
-	out := make([]common.MapStr, 0)
-
-	for _, v := range cs {
-		ms := common.MapStr{}
-
-		oracle.SetSqlValue(m.Logger(), ms, "username", &oracle.StringValue{NullString: v.username})
-		oracle.SetSqlValue(m.Logger(), ms, "machine", &oracle.StringValue{NullString: v.machine})
-		oracle.SetSqlValue(m.Logger(), ms, "cursors.total", &oracle.Int64Value{NullInt64: v.total})
-		oracle.SetSqlValue(m.Logger(), ms, "cursors.max", &oracle.Int64Value{NullInt64: v.max})
-		oracle.SetSqlValue(m.Logger(), ms, "cursors.avg", &oracle.Float64Value{NullFloat64: v.avg})
-
-		out = append(out, ms)
-	}
-
-	return out
-}
-
-func (m *MetricSet) addLibraryData(ls []libraryCache) common.MapStr {
-	out := common.MapStr{}
-
-	for _, v := range ls {
-		if v.name.Valid {
-			oracle.SetSqlValue(m.Logger(), out, v.name.String, &oracle.Float64Value{NullFloat64: v.value})
-		}
-	}
-
-	return out
-}
-
-// addTempFreeSpaceData is specific to the TEMP Tablespace.
-func (m *MetricSet) addBufferCacheRatioData(bs []bufferCacheHitRatio) map[string]common.MapStr {
-	out := make(map[string]common.MapStr)
-
-	for _, bufferCacheHitRatio := range bs {
-		if _, found := out[bufferCacheHitRatio.name.String]; !found {
-			out[bufferCacheHitRatio.name.String] = common.MapStr{}
-		}
-
-		_, _ = out[bufferCacheHitRatio.name.String].Put("buffer_pool", bufferCacheHitRatio.name.String)
-
-		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.buffer.hit.pct", &oracle.Float64Value{NullFloat64: bufferCacheHitRatio.hitRatio})
-		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.get.consistent", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.consistentGets})
-		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.get.db_blocks", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.dbBlockGets})
-		oracle.SetSqlValueWithParentKey(m.Logger(), out, bufferCacheHitRatio.name.String, "cache.physical_reads", &oracle.Int64Value{NullInt64: bufferCacheHitRatio.physicalReads})
-
-	}
-
-	return out
 }
